Continues modular implementation of charts. #483
