+++ src/views/reports/components/terms-per-taxonomy-block.vue
@@ -86,7 +86,7 @@ export default {
                                     trim: true,
                                     hideOverlappingLabels: false
                                 },
-                                tooltip: true
+                                tooltip: { enabled: true }
                             },
                             yaxis: {
                                 title: {
