+++ src/views/reports/pages/reports-list.vue
@@ -80,58 +80,12 @@
                 <metadata-types-block
                         v-if="selectedCollection && selectedCollection != 'default'" />
             </div>
-            <div 
-                    v-if="selectedCollection && selectedCollection != 'default'"
-                    class="column is-full is-two-fifths-desktop">
-                <div 
-                        v-if="!isFetchingMetadata && metadata.totals && metadata.totals.metadata"
-                        :style="{
-                            maxHeight: ((170 + (metadata.totals.metadata.total * 36)) <= 690 ? (170 + (metadata.totals.metadata.total * 36)) : 690) + 'px'
-                        }"
-                        class="postbox metadata-distribution-box">
-                    <apexchart
-                            :height="100 + (metadata.totals.metadata.total * 36)"
-                            :series="metadataDistributionChartSeries"
-                            :options="metadataDistributionChartOptions" />
-                </div>
-                <div 
-                        v-else
-                        style="min-height=740px"
-                        class="skeleton postbox metadata-distribution-box" />
-            </div>
+            <metadata-distribution-block 
+                    v-if="selectedCollection && selectedCollection != 'default'"/>
         </div>
         <div class="columns">
-            <div 
-                    v-if="metadataList != undefined && (selectedCollection && selectedCollection != 'default')"
-                    class="column is-full">
-                <div
-                        style="margin-top: 0px"
-                        class="postbox">
-                    <label>{{ $i18n.get('label_amount_of_items_per_metadatum_value') }}&nbsp;</label>
-                    <select 
-                            v-if="!isFetchingMetadataList"
-                            name="select_metadata"
-                            id="select_metadata"
-                            :placeholder="$i18n.get('label_select_a_metadatum')"
-                            v-model="selectedMetadatum">
-                        <option 
-                                v-for="(metadatum, index) of metadataListArray"
-                                :key="index"
-                                :value="metadatum.id">
-                            {{ metadatum.name }} 
-                        </option>
-                    </select>
-                    <apexchart
-                            v-if="!isFetchingMetadataList && selectedMetadatum"
-                            height="380px"
-                            :series="metadataListChartSeries"
-                            :options="metadataListChartOptions" />
-                    <div 
-                        v-else
-                        style="min-height=380px"
-                        class="skeleton postbox" />
-                </div>
-            </div>
+            <metadata-list-block 
+                    v-if="selectedCollection && selectedCollection != 'default'" />
         </div>
     </div>
 </template>
@@ -151,13 +105,8 @@ export default {
             isFetchingMetadata: false,
             isFetchingMetadataList: false,
             isFetchingActivities: false,
-            collectionsListChartSeries: [],
-            collectionsListChartOptions: {},
             metadataListChartSeries: [],
             metadataListChartOptions: {},
-            metadataDistributionChartSeries: [],
-            metadataDistributionChartOptions: {},
-            metadataDistributionChartHeight: 730,
             activitiesChartSeries: [],
             activitiesChartOptions: {}
         }
@@ -173,7 +122,6 @@ export default {
             collectionsList: 'getCollectionsList',
             activities: 'getActivities',
             stackedBarChartOptions: 'getStackedBarChartOptions',
-            horizontalBarChartOptions: 'getHorizontalBarChartOptions',
             //heatMapChartOptions: 'getHeatMapChartOptions'
         }),
         metadataListArray() {
@@ -233,62 +181,6 @@ export default {
             this.isFetchingMetadata = true;
             this.fetchMetadata({ collectionId: this.selectedCollection })
                 .then(() => {
-
-                    if (this.metadata.distribution) {
-
-                        // Building Metadata Distribution Bar chart
-                        const orderedMetadataDistributions = Object.values(this.metadata.distribution).sort((a, b) => b.fill_percentage - a.fill_percentage );
-                        let metadataDistributionValues = [];
-                        let metadataDistributionValuesInverted = [];
-                        let metadataDistributionLabels = [];
-                        const metadataCount = 100 + (this.metadata.totals.metadata.total * 36);
-
-                        orderedMetadataDistributions.forEach(metadataDistribution => {
-                            metadataDistributionValues.push(parseFloat(metadataDistribution.fill_percentage));
-                            metadataDistributionValuesInverted.push(100.0000 - parseFloat(metadataDistribution.fill_percentage).toFixed(4));
-                            metadataDistributionLabels.push(metadataDistribution.name);
-                        })
-
-                        // Sets first metadatum as the selected one 
-                        if (orderedMetadataDistributions.length)
-                            this.selectedMetadatum = orderedMetadataDistributions[0].id;
-
-                        this.metadataDistributionChartSeries = [
-                            { 
-                                name: this.$i18n.get('label_filled'),
-                                data: metadataDistributionValues
-                            },
-                            { 
-                                name: this.$i18n.get('label_not_filled'),
-                                data: metadataDistributionValuesInverted
-                            }
-                        ];
-                        this.metadataDistributionChartOptions = {
-                            ...this.horizontalBarChartOptions,
-                            ...{
-                                chart: {
-                                    type: 'bar',
-                                    height: metadataCount,
-                                    stacked: true,
-                                    stackType: '100%',
-                                    toolbar: {
-                                        show: true
-                                    },
-                                    zoom: {
-                                        type: 'y',
-                                        enabled: true,
-                                        autoScaleYaxis: true,
-                                    }
-                                },
-                                title: {
-                                    text: this.$i18n.get('label_metadata_fill_distribution')
-                                },
-                                labels: metadataDistributionLabels,
-                                colors: ['#25a189', '#a23939']
-                            }
-                        }
-                    }
-
                     this.isFetchingMetadata = false;
                 })
                 .catch(() => this.isFetchingMetadata = false);
@@ -348,7 +240,7 @@ export default {
                                     trim: true,
                                     hideOverlappingLabels: false
                                 },
-                                tooltip: true
+                                tooltip: { enabled: true }
                             },
                             yaxis: {
                                 title: {
@@ -365,42 +257,7 @@ export default {
         loadMetadataList() {
             this.isFetchingMetadataList = true;
             this.fetchMetadataList({ collectionId: this.collectionId, metadatumId: this.selectedMetadatum })
-                .then(() => {
-                    
-                    // Building Metadata term usage chart
-                    const orderedMetadata = Object.values(this.metadataList).sort((a, b) => b.total_items - a.total_items);
-                    let metadataItemValues = [];
-                    let metadataItemLabels = [];
-
-                    orderedMetadata.forEach(metadataItem => {
-                        metadataItemValues.push(metadataItem.total_items);
-                        metadataItemLabels.push(metadataItem.label);
-                    }); 
-
-                    this.metadataListChartSeries = [
-                        {
-                            name: this.$i18n.get('label_items_with_this_metadum_value'),
-                            data: metadataItemValues
-                        }
-                    ];
-                    
-                    this.metadataListChartOptions = {
-                        ...this.stackedBarChartOptions, 
-                        ...{
-                            title: {},
-                            xaxis: {
-                                type: 'category',
-                                tickPlacement: 'on',
-                                categories: metadataItemLabels,
-                            },
-                            yaxis: {
-                                title: {
-                                    text: this.$i18n.get('label_number_of_items')
-                                }
-                            }
-                        }
-                    }
-                    
+                .then(() => { 
                     this.isFetchingMetadataList = false;
                 })
                 .catch(() => this.isFetchingMetadataList = false);
